Create event processor
