+++ lib/learn_test/cli_event_processor.rb
@@ -1,0 +1,28 @@
+module LearnTest
+  class CliEventProcessor
+    def initialize(learn_profile, lesson_profile, results)
+      @learn_profile  = learn_profile
+      @lesson_profile = lesson_profile
+      @prompter = InterventionPrompter.new(learn_profile, lesson_profile, results)
+    end
+
+    def execute
+      unprocessed_cli_events.each do |event|
+        processed_cli_event!(event)
+        prompter.ask_a_question
+      end
+    end
+
+    private
+
+    attr_reader :lesson_profile, :prompter
+
+    def processed_cli_event!(event)
+      lesson_profile.add_processed_cli_event!(event)
+    end
+
+    def unprocessed_cli_events
+      lesson_profile.unacknowledged_cli_events - lesson_profile.processed_cli_events
+    end
+  end
+end
