+++ lib/learn_test/runner.rb
@@ -3,6 +3,7 @@ require 'yaml'
 module LearnTest
   class Runner
     attr_reader :repo, :options, :results
+
     SERVICE_URL = 'http://ironbroker-v2.flatironschool.com'
 
     def initialize(repo, options = {})
@@ -23,19 +24,7 @@ module LearnTest
       strategy.cleanup unless keep_results?
 
       sync_profiles!
-      trigger_callbacks
-    end
-
-    def prompter
-      LearnTest::InterventionPrompter.new(results, repo, strategy.learn_oauth_token, learn_profile)
-    end
-
-    def trigger_callbacks
-      prompter.execute
-    end
-
-    def learn_profile
-      LearnTest::LearnProfile.new(strategy.learn_oauth_token)
+      process_events!
     end
 
     def files
@@ -52,6 +41,17 @@ module LearnTest
 
     private
 
+    attr_reader :lesson_profile
+
+    def learn_profile
+      LearnTest::LearnProfile.new(strategy.learn_oauth_token)
+    end
+
+    def process_events!
+      event_processor = CliEventProcessor.new(learn_profile, lesson_profile, results)
+      event_processor.execute
+    end
+
     def sync_profiles!
       pid = fork do
         learn_profile.sync!
@@ -60,8 +60,6 @@ module LearnTest
 
       Process.detach(pid)
     end
-
-    attr_reader :lesson_profile
 
     def augment_results!(results)
       if File.exist?("#{FileUtils.pwd}/.learn")
